+++ src/Layout/TaskLayoutOptions.ts
@@ -1,0 +1,19 @@
+import { type TaskLayoutComponent, defaultLayout } from '../TaskLayout';
+
+export class TaskLayoutOptions2 {
+    private visible: { [component: string]: boolean } = {};
+
+    constructor() {
+        defaultLayout.forEach((component) => {
+            this.visible[component] = true;
+        });
+    }
+
+    public isShown(component: TaskLayoutComponent) {
+        return this.visible[component];
+    }
+
+    public hide(component: TaskLayoutComponent) {
+        this.visible[component] = false;
+    }
+}
